refactor: [M3-8801] - Remove unused product icons and rename remaining SVG files to reflect product families
