+++ packages/manager/src/features/ObjectStorage/BucketLanding/StylesBucketIcon.ts
@@ -1,8 +1,8 @@
 import { styled } from '@mui/material/styles';
 
-import BucketIcon from 'src/assets/icons/entityIcons/bucket.svg';
+import StorageIcon from 'src/assets/icons/entityIcons/storage.svg';
 
-const StyledBucketIcon = styled(BucketIcon)(() => ({
+const StyledBucketIcon = styled(StorageIcon)(() => ({
   transform: 'scale(0.80)',
 }));
